data/requirements.txt
loguru~=0.7.3
pandas~=2.3.3
psycopg2~=2.9.11
psycopg2_binary~=2.9.11
PyYAML~=6.0.1
PyYAML~=6.0.3
Requests~=2.32.5
tqdm~=4.66.5
pytest~=8.0.0
pytest-mock~=3.12.0
pytest-cov~=4.1.0